+++ commons-lang/src/main/java/com/hb0730/commons/lang/convert/BasicTypeEnum.java
@@ -52,7 +52,7 @@ public enum BasicTypeEnum {
      * @return 包装类
      */
     public static Class<?> wrap(Class<?> clazz) {
-        if (null == clazz || false == clazz.isPrimitive()) {
+        if (null == clazz || !clazz.isPrimitive()) {
             return clazz;
         }
         Class<?> result = primitiveWrapperMap.get(clazz);