+++ commons-lang/src/test/java/com/hb0730/commons/lang/ClassUtilsTest.java
@@ -1,6 +1,8 @@
 package com.hb0730.commons.lang;
 
-import lombok.*;
+import lombok.Builder;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.Assert;
 import org.junit.Test;