+++ commons-lang/src/test/java/com/hb0730/commons/lang/date/DateUtilsTest.java
@@ -1,58 +1,105 @@
 package com.hb0730.commons.lang.date;
 
+import lombok.extern.slf4j.Slf4j;
+import org.junit.Assert;
 import org.junit.Test;
 
+import java.util.Calendar;
+import java.util.Date;
 import java.util.concurrent.TimeUnit;
 
+@Slf4j
 public class DateUtilsTest {
 
     @Test
     public void nowTest() {
-        System.out.println(DateUtils.now());
+        Date now = DateUtils.now();
+        Assert.assertNotNull(now);
+        log.info(now.toString());
     }
 
     @Test
     public void addTest() {
-        System.out.println(DateUtils.add(DateUtils.now(), 1, TimeUnit.DAYS));
+        Date add = DateUtils.add(DateUtils.now(), 1, TimeUnit.DAYS);
+        Assert.assertNotNull(add);
+        log.info(add.toString());
     }
 
     @Test
     public void addYearsTest() {
-        System.out.println(DateUtils.addYears(DateUtils.now(), 1));
+        Date date = DateUtils.addYears(DateUtils.now(), 1);
+        log.info(date.toString());
     }
 
     @Test
     public void addMonthsTest() {
-        System.out.println(DateUtils.addMonths(DateUtils.now(), 1));
+        Date date = DateUtils.addMonths(DateUtils.now(), 1);
+        log.info(date.toString());
     }
 
     @Test
     public void addWeeksTest() {
-        System.out.println(DateUtils.addWeeks(DateUtils.now(), 1));
+        Date date = DateUtils.addWeeks(DateUtils.now(), 1);
+        log.info(date.toString());
     }
 
     @Test
     public void addDaysTest() {
-        System.out.println(DateUtils.addDays(DateUtils.now(), 1));
+        Date date = DateUtils.addDays(DateUtils.now(), 1);
+        log.info(date.toString());
     }
 
     @Test
     public void addHoursTest() {
-        System.out.println(DateUtils.addHours(DateUtils.now(), 1));
+        Date date = DateUtils.addHours(DateUtils.now(), 1);
+        log.info(date.toString());
     }
 
     @Test
     public void addMinutesTest() {
-        System.out.println(DateUtils.addMinutes(DateUtils.now(), 30));
+        Date date = DateUtils.addMinutes(DateUtils.now(), 30);
+        log.info(date.toString());
     }
 
     @Test
     public void addSecondsTest() {
-        System.out.println(DateUtils.addSeconds(DateUtils.now(), 1800));
+        Date date = DateUtils.addSeconds(DateUtils.now(), 1800);
+        log.info(date.toString());
     }
 
     @Test
     public void addMillisecondsTest() {
-        System.out.println(DateUtils.addMilliseconds(DateUtils.now(), 30 * 60 * 1000));
+        Date date = DateUtils.addMilliseconds(DateUtils.now(), 30 * 60 * 1000);
+        log.info(date.toString());
+    }
+
+    @Test
+    public void betweenTest() {
+        long between = DateUtils.between(DateUtils.now(), DateUtils.addDays(DateUtils.now(), 1), DateMsUnit.DAY);
+        Assert.assertNotNull(between);
+        log.info(String.valueOf(between));
+    }
+
+    @Test
+    public void testBetweenTest() {
+        long between = DateUtils.between(DateUtils.now(), DateUtils.addDays(DateUtils.now(), 1), DateMsUnit.DAY, false);
+        Assert.assertNotNull(between);
+        log.info(String.valueOf(between));
+        between = DateUtils.between(DateUtils.addDays(DateUtils.now(), 1), DateUtils.now(), DateMsUnit.DAY, true);
+        Assert.assertNotNull(between);
+        log.info(String.valueOf(between));
+    }
+
+    @Test
+    public void getServerStartDateTest() {
+        Date date = DateUtils.getServerStartDate();
+        Assert.assertNotNull(date);
+        log.info(date.toString());
+    }
+
+    @Test
+    public void convertToTest() {
+        Calendar calendar = DateUtils.convertTo(DateUtils.now());
+        Assert.assertNotNull(calendar);
     }
 }